Update teams-powershell-release-notes.md
New-Team no longer returns a Teams object but an object with a single 'TargetResourceLocation' string property. This will be a breaking change for many users. This PR includes text to warn about the change and links to the New-Team documentation. Note, there is an outstanding PR to update the documentation for New-Team that should also be approved to help with this: https://github.com/MicrosoftDocs/office-docs-powershell/pull/8220