Update create-a-phone-system-call-queue.md
Please make it more clear that it is a requirement that the admin completing the task to add a channel to a Call Queue is required to be a member or owner. Requiring an admin to be a member or owner of a team that they are administering is not normal practice.