Acrolinx fixes for minimum score...
(1) Removed "th" from dates; that yielded almost enough of an improvement to get to 80.
(2) Put the version numbers that include ".product" inside of backticks, marking these as code, not language; with that, the score was probably above 80.
(3) Put all version numbers inside of backticks; this didn't help the Acrolinx score, but it made presentation consistent.