Rearranged the 'notes' on Dimensions
A lot of important information is included in the notes section after the dimensions table, but it's so far down that customers can often miss it. We are relocating this information to the top of the dimensions section as we get a lot of questions about understanding how range values work.