Update assign-role-permissions.md to clarify assignment instructions.
Users are confused by the absence of **Manage team** when selecting it from a pinned team. This stems from the fact that teams themselves are not pinne. Rather, only channels are pinned. The revised documentation clarifies this by instructing the user to choose the team name under *Your teams* rather than under *pinned*. In addition, the ellipses (**...**) are parenthetically added to **More options** for further clarification.