Updating definitions for 'Poor percentage'
Our Poor Percentage calculations are based on classified streams; that is, streams where enough data is available to determine with sufficient confidence if a call actually was good or poor. Because unclassified streams do not have enough information -- such as low duration or packet utilization -- to make this determination, we do not include these in our measurements. This was unclear in previous versions of the documentation.
-------
cc: @SerdarSoysal