Update browser-support.md
Teams client hardware requirements article mentions that Teams web is not supported on Windows 10 LTSC but in this article under browser support we dont call it explicitly which creates confusion for customers when referring to both Teams client  requirement article. Hence making a change in browser section to show that Teams web on any browser on Windows 10 LTSC is not supported.