Update manage-external-access.md
These updates are in preparation for a user-level feature (1:1 TFW-TFW Accept/Block) that rolls out in June (estimation). They should not be published until the feature is live. This gives us time to iterate on the contents. Updates to Admin docs were requested by Privacy + Security. Happy to also evaluate if this is the incorrect section.