Update archive-or-delete-a-team.md
I would suggest changing the language in step 1  of "Archive a team" and "Make an archived team active," from "Teams" to "Manage teams" to differentiate between the dropdown in the menu and the page selector. 

In step 2, the current action to "Select a team by clicking the team name" opens the team details page. The option in step 3 to Archive is not present on that page. The behavior instead seems to be adding a checkmark in the first column, then the Archive (un Unarchive) button becomes available in the menu/ribbon.

In step 3 of "Make an archived team active," the copy is "Unarchive" (no longer "Restore").

Thanks!