Update teams-app-permission-policies.md
Although there is a statement in the published version that third-party apps are not available in GCCH and DoD, the instructions still reference those environments making it appear that third-party apps are available in them.  The only reference to Microsoft 365 Government Community Cloud High (GCCH) and Microsoft 365 Department of Defense (DoD) should be that note.  The instructions under "Manage org-wide app settings for Microsoft 365 Government" are only valid for the Microsoft 365 Government Community Cloud (GCC) environment.