Correct inconsistency in the Engagement report
The Engagement report default value is Off. This is mentioned in the notes of [View and download meeting attendance reports in Teams](https://support.microsoft.com/en-us/office/view-and-download-meeting-attendance-reports-in-teams-ae7cf170-530c-47d3-84c1-3aedac74d310) and in [Microsoft Teams meeting attendance report](https://docs.microsoft.com/en-us/microsoftteams/teams-analytics-and-reports/meeting-attendance-report). The option is turned off when a new Teams meeting policy is created from Teams admin portal, but Enabled when a new Teams meeting policy is created in PowerShell.