SBA requires ICMP otherwise it will not work.
ICMP to www.microsoft.com is required as part of the firewall requirements.
If ICMP is denied on the firewall then the SBA will not function and there is a log  every 30 seconds in the sba.log  "No Internet Connectivity. Waiting 30 seconds."  if you then allow ICMP the SBA works.
If you start and stop the Teams SBA service, then the ICMP ping to www.microsoft.com also starts/stops
Without ICMP the SBA will not function!