IP phone policies section
Since the license change, we are now automatically assigning and enforcing the CommonAreaPhoneSignIn mode for any phone that is licensed with a CAP license.  It is not permitted to change the IP Phone policy to anything other than CommonAreaPhoneSignIn.   specified here.  https://docs.microsoft.com/en-us/microsoftteams/devices/teams-android-devices-user-interface#override-automatic-user-interface-detection 
I think this section should be modified to let customers know that they can no longer change the SignIn mode for a Common Area phone that is licensed with a CAP license.