Update teams-security-best-practices-for-safer-messaging.md
Line 34: Just as with email,> As with email
Line 40: Again, use your best judgment for your own, and the safety of other users, as well as your customers. > Again, use your best judgment for your own, other users', and your customers' safety.
Line 41:If selecting a link lands you on a safe links page where you are blocked from navigation, don't try to circumvent that page (the same goes for any clicked attachment that may end in a red, blocking safe attachments page). If you know and trust the target site, report the problem to Microsoft and to whomever directed you there. But there are many reasons to land on a blocking page, and red flags should be considered, rather than circumnavigated. > If selecting a link lands you on a safe links page where you are blocked from navigation, don't try to circumvent that page. The same goes for selecting any attachment that may end in a red, blocking attachments safe page. If you know and trust the target site, report the problem to Microsoft and to whomever directed you there. Many reasons exist to land on a blocking page, so consider the red flags behind those reasons, rather than circumnavigate the blocking page.
Line 42: This is also a risk to personal safety.  Details as simple as your birthday can be leveraged by attackers. > Providing personal information is also a risk to personal safety. Attackers can leverage details as simple as your birthday.
Line 45: It is important to maintain your vigilance and not take your security, and the security of those you work with lightly, or for granted. > Maintaining your vigilance is important. Don't take your security or the security of those you work with lightly or for granted. You should always feel empowered to protect yourself, your peers, and your customers by verifying before you trust.
Line 47: Finally, it's also critical to recognize that everyone makes mistakes. Users can be prone to clicking when in a hurry, for example, or when tired. Admins should make sure the resources to escalate problem link clicks and other security incidents to are clearly known to users in their organization, so that, if a mistake is made, the user has recourse, and further security actions can be taken. > Finally, recognizing that everyone makes mistakes is critical. Users are prone to clicking when in a hurry or tired, for example. Admins must provide users in their organization with resources to escalate problematic link clicks and other security incidents and make those resources clearly known so that if a user makes a mistake, the user has recourse, and further security actions are available.