Create Disable Legacy Control Panel (CSCP).md
As part of the GB cert compliance effort, for MS to sell products in China region, we need to make sure all GB characters are valid in our product. 

Our old control panel has an issue with GB characters and since we have a modern control panel, the script in this document will disable  old control panel making us compliant.