Update private-channels.md
Updated the document according to how this currently works.
Team owners can't see the names of private channels in their team that they are not members of or that they don't own. A private channel owner or team owner who is a member of the private channel can delete the private channel in the team. (A deleted private channel can be restored within 30 days after it's deleted). Team owners can't see the channels, files in a private channel, or the conversations and member list of a private channel unless they're members of that private channel.