Update presence-admins.md (default duration for invisibility)
When you set your presence to "Appear offline" (i.e. invisible) without setting a duration, the end date is actually set to DateTime.UtcMaxValue. This is correct and has always been the case, but our documentation mistakenly said that it's set for 7 days, which caused some confusion today.

This PR updates the documentation to reflect what's happening in code.