Replace 404'ed link in new-teams-desktop-admin.md
Replaced a 404'ed link with what appears to be the intended replacement destination

Old:
- /office/try-the-new-microsoft-teams-2d4a0c96-fa52-43f8-a006-4bfbc62cf6c5
-https://learn.microsoft.com/en-us/office/try-the-new-microsoft-teams-2d4a0c96-fa52-43f8-a006-4bfbc62cf6c5

New:
-/office/switch-to-the-new-microsoft-teams-2d4a0c96-fa52-43f8-a006-4bfbc62cf6c5
-https://support.microsoft.com/en-us/office/switch-to-the-new-microsoft-teams-2d4a0c96-fa52-43f8-a006-4bfbc62cf6c5

I'm not sure of the above syntax, i.e. full path versus abbreviated path, and/or if the abbreviated path will work since this page is on learn.microsoft and the destination is on support.microsoft.