Fix redirection entries
The problem is that PR 15660 added new entries between line 5196 and what was line 5267, and PR 15655 adds new entries between lines 5246 and what was line 5257. Also, the values for redirect_url are based on the URLs of the targeted pages, not the file paths in the repo. So, any page in the repo in the Teams folder should have a redirect_url value that starts with /microsoftteams/.

After fixing this, I've also restored proper indentation of the whole file.